+++ facade/src/main/scala/react/semanticui/package.scala
@@ -92,7 +92,8 @@ package object semanticui
     with Colors
     with Sizes
     with TextAlignment
-    with VerticalAlignment {
+    with VerticalAlignment
+    with Transitions {
   type SemanticShortHandItem[T] = VdomNode | T
   type SemanticShortHandContent = VdomNode
   type TabIndex                 = Double | String
@@ -103,6 +104,7 @@ package object semanticui
   val sizes             = SemanticSize
   val textalignment     = SemanticTextAlignment
   val verticalalignment = SemanticVerticalAlignment
+  val transitions       = SemanticTransition
 
   implicit class HandContent2Content(val c: SemanticShortHandContent) extends AnyVal {
     def toRaw: raw.SemanticShorthandContent = c.rawNode